fix performance

+++ velox/experimental/stateful/agg/GroupAggregator.h
@@ -61,7 +61,7 @@ class GroupAggregator : public exec::Operator, public KeyedProcessFunction {
 
  private:
   std::unique_ptr<AggsHandleFunction> aggsFunction_;
-  std::shared_ptr<ValueState<uint32_t, int64_t, RowVectorPtr>> accState_;
+  std::shared_ptr<ValueState<int64_t, int64_t, RowVectorPtr>> accState_;
   int64_t stateRetentionTime_;
   bool generateUpdateBefore_;
 };
